+++ analyzers/tests/SonarAnalyzer.Test/TestCases/TestMethodShouldContainAssertion.MsTest.cs
@@ -7,7 +7,6 @@
     using NSubstitute;
     using Microsoft.VisualStudio.TestTools.UnitTesting;
     using Shouldly;
-    using Moq;
 
     using static Microsoft.VisualStudio.TestTools.UnitTesting.Assert;
 
@@ -355,74 +354,6 @@
         {
             int[] array = { 1, 2, 3 };
             array.ShouldAllBe(x => x > 0);
-        }
-    }
-
-    public delegate void CalculatorEvent(int i, bool b);
-    internal interface ICalculator
-    {
-        int Property { get; set; }
-        event CalculatorEvent Event;
-        int Add(int a, int b);
-    }
-
-    [TestClass]
-    public class MoqVerifyTests
-    {
-        [TestMethod]
-        public void MoqVerify()
-        {
-            var mock = new Mock<ICalculator>();
-
-            mock.Verify(x => x.Add(It.IsAny<int>(), It.IsAny<int>()), Times.Never());
-        }
-
-        [TestMethod]
-        public void MoqVerifyAdd()
-        {
-            var mock = new Mock<ICalculator>();
-
-            mock.VerifyAdd(x => x.Event += It.IsAny<CalculatorEvent>());
-        }
-
-        [TestMethod]
-        public void MoqVerifyRemove()
-        {
-            var mock = new Mock<ICalculator>();
-
-            mock.VerifyRemove(x => x.Event -= It.IsAny<CalculatorEvent>(), Times.AtMostOnce());
-        }
-
-        [TestMethod]
-        public void MoqVerifySet()
-        {
-            var mock = new Mock<ICalculator>();
-
-            mock.VerifySet(x => x.Property = It.IsAny<int>());
-        }
-
-        [TestMethod]
-        public void MoqVerifyGet()
-        {
-            var mock = new Mock<ICalculator>();
-
-            mock.VerifyGet(x => x.Property);
-        }
-
-        [TestMethod]
-        public void MoqVerifyNoOtherCalls()
-        {
-            var mock = new Mock<ICalculator>();
-
-            mock.VerifyNoOtherCalls();
-        }
-
-        [TestMethod]
-        public void MoqVerifyAll()
-        {
-            var mock = new Mock<ICalculator>();
-
-            mock.VerifyAll();
         }
     }
